extract prop to params to utils
